# Simplify Parsing Logic for a performance boost