Removes the stack accumulator from the `unpack` function and simplifies the collection logic so that each loop of the collection matches the binary in the function head.

This makes unpacking have a 20-50% better memory usage and 0-15% increase in speed.

# Benchmarks

## 100 byte 2 dimensional list of integers

```
Name                                                          ips        average  deviation         median         99th %
Elixir.MsgpaxHansonkd - Complexity: 2 - Size: small      354.28 K        2.82 μs  ±1809.37%           2 μs           6 μs
Elixir.Msgpax - Complexity: 2 - Size: small              353.73 K        2.83 μs  ±1600.46%           2 μs           7 μs

Comparison: 
Elixir.MsgpaxHansonkd - Complexity: 2 - Size: small      354.28 K
Elixir.Msgpax - Complexity: 2 - Size: small              353.73 K - 1.00x slower

Memory usage statistics:

Name                                                   Memory usage
Elixir.MsgpaxHansonkd - Complexity: 2 - Size: small         1.04 KB
Elixir.Msgpax - Complexity: 2 - Size: small                 1.52 KB - 1.47x memory usage

**All measurements for memory usage were the same**
```

## 1kb binary with 5 dimensional list of integers
```
Name                                                          ips        average  deviation         median         99th %
Elixir.MsgpaxHansonkd - Complexity: 5 - Size: large         41.58       24.05 ms     ±7.96%       23.87 ms       30.46 ms
Elixir.Msgpax - Complexity: 5 - Size: large                 36.58       27.34 ms     ±6.71%       26.99 ms       36.04 ms

Comparison: 
Elixir.MsgpaxHansonkd - Complexity: 5 - Size: large         41.58
Elixir.Msgpax - Complexity: 5 - Size: large                 36.58 - 1.14x slower

Memory usage statistics:

Name                                                   Memory usage
Elixir.MsgpaxHansonkd - Complexity: 5 - Size: large         6.79 MB
Elixir.Msgpax - Complexity: 5 - Size: large                 8.04 MB - 1.18x memory usage
```

## 1kb binary with 1 dimensional map of string keys
```
Name                                                          ips        average  deviation         median         99th %
Elixir.MsgpaxHansonkd - Complexity: 1 - Size: large       61.64 K       16.22 μs    ±80.63%          15 μs          32 μs
Elixir.Msgpax - Complexity: 1 - Size: large               48.93 K       20.44 μs    ±65.72%          19 μs          40 μs

Comparison: 
Elixir.MsgpaxHansonkd - Complexity: 1 - Size: large       61.64 K
Elixir.Msgpax - Complexity: 1 - Size: large               48.93 K - 1.26x slower

Memory usage statistics:

Name                                                   Memory usage
Elixir.MsgpaxHansonkd - Complexity: 1 - Size: large         9.76 KB
Elixir.Msgpax - Complexity: 1 - Size: large                14.59 KB - 1.49x memory usage
```

## 1.3mb binary with 5 dimensional map 
```
Name                                                          ips        average  deviation         median         99th %
Elixir.MsgpaxHansonkd - Complexity: 5 - Size: large         17.65       56.64 ms     ±7.71%       55.40 ms       73.41 ms
Elixir.Msgpax - Complexity: 5 - Size: large                 12.82       78.03 ms     ±3.21%       77.74 ms       87.39 ms

Comparison: 
Elixir.MsgpaxHansonkd - Complexity: 5 - Size: large         17.65
Elixir.Msgpax - Complexity: 5 - Size: large                 12.82 - 1.38x slower

Memory usage statistics:

Name                                                   Memory usage
Elixir.MsgpaxHansonkd - Complexity: 5 - Size: large        17.96 MB
Elixir.Msgpax - Complexity: 5 - Size: large                24.50 MB - 1.36x memory usage
```