@lexmag 2020-12-12
@hansonkd thanks a lot for the PR (and your blog post about Msgpax).
I did consider to specialize the collection loops when I was doing #30 and decided not to do it due to:
1. it drastically affects compilation time and beam binary size

**OTP 20**, Elixir 1.6
master: 1.5 sec, 7.5 KB
this PR: 3.3 sec, 116 KB

**OTP 23**, Elixir 1.11
master: 2.4 sec, 8.1 KB
this PR: 7.8 sec, 34 KB

With OTP 20 it gives a too large beam binary, while with OTP 23 (was not available back then) the beam binary does not grow that crazy but it takes uncomfortably long time to compile.

2. it complicates the code but the benefits it brings I tend to think mostly to some specialised use cases
3. it seems master has more stable performance (less deviation)

Perhaps there is a middle ground in terms of specialzing the collection loops (I tried to remove `unpack_map_value` inlining and it improves compilation time and beam binary size) but that requires further investigation and more benchmarking.

What are your thoughts on this?