@lexmag 2021-05-07
I'm closing the PR for now, but feel free to reopen and we can discuss how to proceed with potential optimizations.